move extraction logic to seperate common package
