+++ receiver/oracledbreceiver/scraper.go
@@ -14,7 +14,6 @@ import (
 	"math"
 	"net"
 	"os"
-	"regexp"
 	"sort"
 	"strconv"
 	"strings"
@@ -32,6 +31,7 @@ import (
 	"go.uber.org/multierr"
 	"go.uber.org/zap"
 
+	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/common/sqlcomments"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/oracledbreceiver/internal/metadata"
 )
 
@@ -110,11 +110,6 @@ var (
 	oracleQueryMetricsSQL string
 	//go:embed templates/oracleQueryPlanSql.tmpl
 	oracleQueryPlanDataSQL string
-
-	// leadingBlockCommentRegex matches one or more leading /* */ block comments
-	leadingBlockCommentRegex = regexp.MustCompile(`^\s*(/\*.*?\*/\s*)+`)
-	// commentContentRegex extracts content between /* and */ delimiters
-	commentContentRegex = regexp.MustCompile(`/\*(.*?)\*/`)
 )
 
 type dbProviderFunc func() (*sql.DB, error)
@@ -148,83 +143,6 @@ type oracleScraper struct {
 	querySampleCfg             QuerySample
 	serviceInstanceID          string
 	lastExecutionTimestamp     time.Time
-}
-
-// extractAndFilterComments extracts leading /* */ block comments from SQL,
-// parses them as key=value pairs, and returns only allowed keys.
-// Returns comma-separated filtered pairs, or empty string if no allowed keys found.
-// Format: "key1=value1,key2=value2"
-func extractAndFilterComments(sqlText string, allowedKeys []string) string {
-	// Early exit: if no allowed keys, return empty immediately (secure by default)
-	if len(allowedKeys) == 0 {
-		return ""
-	}
-
-	// Extract leading block comments using regex
-	matches := leadingBlockCommentRegex.FindString(sqlText)
-	if matches == "" {
-		return ""
-	}
-
-	// Strip /* and */ delimiters from all comments
-	// Match each individual comment block
-	commentMatches := commentContentRegex.FindAllStringSubmatch(matches, -1)
-	if len(commentMatches) == 0 {
-		return ""
-	}
-
-	// Concatenate all comment contents
-	var allComments strings.Builder
-	for i, match := range commentMatches {
-		if len(match) > 1 {
-			if i > 0 {
-				allComments.WriteString(",")
-			}
-			allComments.WriteString(strings.TrimSpace(match[1]))
-		}
-	}
-
-	commentContent := allComments.String()
-	if commentContent == "" {
-		return ""
-	}
-
-	// Parse key=value pairs and filter by allowed keys
-	pairs := strings.Split(commentContent, ",")
-	var filteredPairs []string
-	seenKeys := make(map[string]bool)
-
-	// Create a set of allowed keys for O(1) lookup
-	allowedSet := make(map[string]bool)
-	for _, key := range allowedKeys {
-		allowedSet[key] = true
-	}
-
-	for _, pair := range pairs {
-		pair = strings.TrimSpace(pair)
-		if pair == "" {
-			continue
-		}
-
-		// Split by first = only
-		parts := strings.SplitN(pair, "=", 2)
-		if len(parts) != 2 {
-			// Malformed pair, skip it
-			continue
-		}
-
-		key := strings.TrimSpace(parts[0])
-		value := strings.TrimSpace(parts[1])
-
-		// Check if key is allowed and not already seen (use first occurrence)
-		if allowedSet[key] && !seenKeys[key] {
-			seenKeys[key] = true
-			filteredPairs = append(filteredPairs, key+"="+value)
-		}
-	}
-
-	// Join filtered pairs with comma (no space)
-	return strings.Join(filteredPairs, ",")
 }
 
 func newScraper(metricsBuilder *metadata.MetricsBuilder, metricsBuilderConfig metadata.MetricsBuilderConfig, scrapeCfg scraperhelper.ControllerConfig, logger *zap.Logger, providerFunc dbProviderFunc, clientProviderFunc clientProviderFunc, instanceName, hostName string) (scraper.Metrics, error) {
@@ -693,7 +611,7 @@ func (s *oracleScraper) collectTopNMetricData(ctx context.Context, logs plog.Log
 			}
 
 			// Extract and filter comments from original SQL before obfuscation
-			queryComments := extractAndFilterComments(row[sqlTextAttr], s.topQueryCollectCfg.AllowedCommentKeys)
+			queryComments := sqlcomments.ExtractAndFilterComments(row[sqlTextAttr], s.topQueryCollectCfg.AllowedCommentKeys)
 
 			hit := queryMetricCacheHit{
 				sqlID:         row[sqlIDAttr],
@@ -878,7 +796,7 @@ func (s *oracleScraper) collectQuerySamples(ctx context.Context, logs plog.Logs)
 		})
 
 		// Extract and filter query comments from original SQL (before obfuscation)
-		queryComments := extractAndFilterComments(row[sqlText], s.querySampleCfg.AllowedCommentKeys)
+		queryComments := sqlcomments.ExtractAndFilterComments(row[sqlText], s.querySampleCfg.AllowedCommentKeys)
 
 		s.lb.RecordDbServerQuerySampleEvent(queryContext, timestamp, obfuscatedSQL, dbSystemNameVal, row[username], row[serviceName], row[hostName],
 			clientPort, row[hostName], clientPort, queryPlanHashVal, row[sqlID], row[sqlChildNumber], row[childAddress], row[sid], row[serialNumber], row[process],
